# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Photo Viewer API",
        "contact": {},
        "version": "dev"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/collection/{collection_uuid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Collection info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection UUID",
                        "name": "collection_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Access link secret (for protected collections)",
                        "name": "collection_access_secret",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.CollectionInfo"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "collection is not permitted",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "collection not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Delete a collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection UUID",
                        "name": "collection_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token is empty / invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "not permitted (not the owner)",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/photo/{photo_uuid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Photo info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Photo UUID",
                        "name": "photo_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Access link secret (for protected photos)",
                        "name": "photo_access_secret",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.PhotoInfo"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "photo is not permitted",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "photo not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Delete a photo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Photo UUID",
                        "name": "photo_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token is empty / invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "invalid authorization (not the owner)",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Update photo metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Photo UUID",
                        "name": "photo_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JSON: title (≤50), description (≤200), created_at, took_at, access_level (private|protected|public)",
                        "name": "metadata",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/update.Response"
                        }
                    },
                    "400": {
                        "description": "invalid metadata / validation / invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token is empty / invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "photo not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/photos": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "List photos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owner login; without the param — all public photos",
                        "name": "owner_login",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.PhotoInfo"
                            }
                        }
                    },
                    "404": {
                        "description": "owner not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Total body size limit is 10 MB.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Upload a photo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JSON: title (≤50), description (≤200), created_at, took_at, access_level (private|protected|public), tag_uuids []",
                        "name": "metadata",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Photo file",
                        "name": "photo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/upload.Response"
                        }
                    },
                    "400": {
                        "description": "invalid metadata / validation / invalid photo file / file too big / tag does not exist",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token is empty / invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "quota is not enough",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.AccessModifier": {
            "type": "string",
            "enum": [
                "private",
                "protected",
                "public"
            ],
            "x-enum-varnames": [
                "AccessModifierPrivate",
                "AccessModifierProtected",
                "AccessModifierPublic"
            ]
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Some error description"
                },
                "status": {
                    "type": "string",
                    "example": "Ok"
                }
            }
        },
        "service.CollectionInfo": {
            "type": "object",
            "required": [
                "access_level"
            ],
            "properties": {
                "access_level": {
                    "$ref": "#/definitions/entity.AccessModifier"
                },
                "collection_uuid": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 200
                },
                "owner_login": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.PhotoSmallInfo"
                    }
                },
                "title": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
...
        "service.PhotoInfo": {
            "type": "object",
            "required": [
                "access_level"
            ],
            "properties": {
                "access_key": {
                    "type": "string"
                },
                "access_level": {
                    "$ref": "#/definitions/entity.AccessModifier"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 200
                },
                "owner_login": {
                    "type": "string"
                },
                "photo_uuid": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.TagSmallInfo"
                    }
                },
                "title": {
                    "type": "string",
                    "maxLength": 50
                },
                "took_at": {
                    "type": "string"
                }
            }
        },
        "service.PhotoSmallInfo": {
            "type": "object",
            "properties": {
                "access_key": {
                    "type": "string"
                },
                "photo_uuid": {
                    "type": "string"
                }
            }
        },
...
        "service.TagSmallInfo": {
            "type": "object",
            "properties": {
                "tag_name": {
                    "type": "string"
                },
                "tag_uuid": {
                    "type": "string"
                }
            }
        },
...
        "update.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Some error description"
                },
                "photo_uuid": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "Ok"
                }
            }
        },
...
        "upload.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Some error description"
                },
                "photo_uuid": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "Ok"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: \"Bearer \u003caccess_token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}