# internal/http-server/handlers/url/view/view-photo-info.go
package view

import (
	"context"
	"errors"
	"log/slog"
	"net/http"
	"photo-viewer-server/internal/lib/api/response"
	"photo-viewer-server/internal/service"
	"photo-viewer-server/internal/storage"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/render"
	"github.com/google/uuid"
)

// ViewPhotoInfo returns photo metadata by UUID.
//
//	@Summary		Photo info
//	@Tags				photos
//	@Produce		json
//	@Param			photo_uuid	path		string	true	"Photo UUID"
//	@Param			photo_access_secret	query	string	false	"Access link secret (for protected photos)"
//	@Success		200			{object}	service.PhotoInfo
//	@Failure		400			{object}	response.Response	"invalid request"
//	@Failure		403			{object}	response.Response	"photo is not permitted"
//	@Failure		404			{object}	response.Response	"photo not found"
//	@Failure		500			{object}	response.Response	"internal error"
//	@Router			/photo/{photo_uuid} [get]
func ViewPhotoInfo(lg *slog.Logger, photoService *service.PhotoService) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		log := lg.With(
			slog.String("op", "handlers.view.ViewPhotos"),
			slog.String("request_id", middleware.GetReqID(r.Context())),
		)

		photoIdStr := chi.URLParam(r, "photo_uuid")
		if photoIdStr == "" {
			log.Info("photo id param is empty")

			render.Status(r, http.StatusBadRequest)
			render.JSON(w, r, response.Error("photo id is empty"))
			return
		}

		accessKey := r.URL.Query().Get("photo_access_secret")
		r.WithContext(context.WithValue(r.Context(), "photo_access_secret", accessKey))

		photoUuid, err := uuid.Parse(photoIdStr)
		if err != nil {
			log.Error("failed to convert photo uuid str to uuid", slog.String("photo_id_str", photoIdStr))

			render.Status(r, http.StatusBadRequest)
			render.JSON(w, r, response.Error("invalid request"))
			return
		}

		photoInfo, err := photoService.GetPhotoInfo(r.Context(), photoUuid)
		if err != nil {
			if errors.Is(err, storage.ErrPhotoNotFound) {
				log.Info("photo not found", slog.Any("photo_uuid", photoUuid))

				render.Status(r, http.StatusNotFound)
				render.JSON(w, r, response.Error("photo not found"))
				return

			} else if errors.Is(err, service.ErrPhotoIsNotPermitted) {
				render.Status(r, http.StatusForbidden)
				render.JSON(w, r, response.Error("photo is not permitted"))
				return
			}

			render.Status(r, http.StatusInternalServerError)
			render.JSON(w, r, response.Error("internal error"))
			return
		}

		log.Info("photo found", slog.Any("photo_uuid", photoUuid))
		render.Status(r, http.StatusOK)
		render.JSON(w, r, photoInfo)
	}
}